requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-storages==1.14.4
django-filter==24.3
psycopg2-binary==2.9.10
boto3==1.36.0
razorpay==1.4.2
Pillow==11.1.0
gunicorn==23.0.0
whitenoise==6.8.2
requests==2.32.3
mutagen==1.47.0
celery==5.4.0
redis==5.2.1
